refactor: EventStore api - remove currentFilter
